GitHub Actions pipeline update

+++ .build/pack.mjs
@@ -1,11 +1,12 @@
 import path from "path";
 import fs from "fs";
 import {compilePack} from "@foundryvtt/foundryvtt-cli";
+import {outputFileSync} from "fs-extra/esm";
 
-const inputPath = "../src/macros";
-const jsonPath = "../src/packs/macros";
-const outputPath = "../packs/macros";
-const macrosData = JSON.parse(fs.readFileSync("macros.data.json"));
+const inputPath = "src/macros";
+const jsonPath = "src/packs/macros";
+const outputPath = "packs/macros";
+const macrosData = JSON.parse(fs.readFileSync(".build/macros.data.json"));
 
 function transformData(macro) {
   const data = fs.readFileSync(path.join(inputPath, macro.codeFile), "utf8");
@@ -28,8 +29,10 @@ function transformData(macro) {
   return macro;
 }
 
-for (let filepath of fs.readdirSync(jsonPath)) {
-  fs.unlinkSync(path.join(jsonPath, filepath));
+if (fs.existsSync(jsonPath)) {
+  for (let filepath of fs.readdirSync(jsonPath)) {
+    fs.unlinkSync(path.join(jsonPath, filepath));
+  }
 }
 for (let filepath of fs.readdirSync(inputPath, {recursive: true, withFileTypes: true})) {
   if (!filepath.isFile() || !filepath.name.endsWith(".js")) {
@@ -38,7 +41,7 @@ for (let filepath of fs.readdirSync(inputPath, {recursive: true, withFileTypes: 
   filepath = filepath.name;
   let macro = macrosData.macros.find((m) => m.codeFile === filepath);
   if (!macro) {
-    Utility.error(`No entry for ${filepath}`);
+    console.error(`No entry for ${filepath}`);
     continue;
   }
 
@@ -46,8 +49,9 @@ for (let filepath of fs.readdirSync(inputPath, {recursive: true, withFileTypes: 
   let fileData = macrosData.common;
   fileData = Object.assign(fileData, macro);
   fileData = transformData(fileData);
-  fs.writeFileSync(path.join(jsonPath, fileName), JSON.stringify(fileData, null, 2) + "\n", "utf8");
+  outputFileSync(path.join(jsonPath, fileName), JSON.stringify(fileData, null, 2) + "\n", "utf8");
 }
+console.log("Pack data transformation complete.");
 compilePack(jsonPath, outputPath, {log: false})
-  .then(() => Utility.log("Pack compilation complete."))
-  .catch((err) => Utility.log(err.message));
+  .then(() => console.log("Pack compilation complete."))
+  .catch((err) => console.log(err.message));
